First stages of rewriting fresh to be cleaner, better documented, allow for more customization (build tags), etc.
The original code has basically no documentation, and is quite messy (see reading from "settings" variable for config info) so this is going to be quite a project

I looked, and tried, air, but it is not as performant.  And the code just has another layer of messiness on it that is tough to work with, at least fresh is more "OG".