Continue rewriting and cleaning up.  Kind of making a mess of things while testing, rewriting, etc.
Use newer WalkDirFunc instead of WalkDir...apparently is is more performance.

Use non-custom error for returning skipped directories from WalkDirFunc.  Why use a custom error when go already has an error for this use case.

Create a WalkDirFunc instead of using anonymous function.  Sure, code is a bit more spread out, but I think it is cleaner and a bit easier to follow.

Default config is the build-in config, not reading from a file or writing default config to file. This removes the need to have a config file, or really for a config file to be written to disk upon first use, for no good reason.