Better description of each config file field, rename config file fields, add -tags flag support.
Better describe each config file field, what it is for, what it does, etc. Previous explanations were kept short because they were a single line, but this led to confusion about what each field was for.

Rename fields from "GoBuild..." to just "Go..." to make field names shorter and since some fields are used for `go build` and `go run`.

Add support for -tags flag to fresher that passes flags to `go build -tags`. This way users don't have to have a config file to set -tags or have to modify a config file each time they want to change -tags.  Anything provided to -tags overrides the config file GoTags field.

Check that any extensions to not rebuild on are listed in extensions to watch. It doesn't make sense to not rebuild on an extension that isn't already watched!