A lot of clean up to help with duplicate file change events, some other bugs, and clearer logging.
Sometimes, mostly on Windows, a duplicate file change event occurs.  This has to do with text editor and OS... The file isn't changed in place upon a save, it is saved to a new file then renamed to the old file; this is very low level stuff.  Anyway, a duplicate "change" event occurs which can cause rebuilds to happen twice for no necessary reason.  The NewTimer helps "debounce" these events by waiting for the duplicate and only sending the event on the second/last occurance.

Change how events chan works.  Send the underlying fsnotify event on the channel, not just the file that was changed.  This wasy we have access to the change type elsewhere (mostly for logging).

Set the events chan to have storage for 1 event.  There isn't a need to store more than 1 event since we block on rebuilding anyway.  `fresh` didn't really use multiple slots in the channel either since all messages were flushed when any one message was read anyway.

Have to clean up some logging and put some stuff behind the VerboseLogging flag.  Added a bit of logging to figure out bugs so now they need to be cleaned up.